Tried to upload a package to Maven
I even managed to script the signing part without using Maven or Ant.
But this stuff is really weird in the Java ecosystem. Compare this to
NuGet.

In the end I am defeated. I didn't manage to upload it to the Central
Maven Directory or any other place. I guess this is only possible with
Maven or Ant.